+++ CHANGELOG.md
@@ -9,6 +9,7 @@ overall surface area of the app has been reduced and more opinions have
 been made.
 
 - Renamed `Microcosm::seed` to `Microcosm::push`
+- Renamed `Microcosm::get` to `Microcosm::pull`
 - Removed `Microcosm::has`
 - Removed `Microcosm::getInitialState`. the `Store` API still provides
   this function, however it is the expectation of the system that